requirements.txt
streamlit>=1.24.0
pandas>=1.5.3
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
PyYAML>=6.0 